+++ packages/react/src/components/HelpLink/src/helpLink.stories.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { ComponentMeta, ComponentStory } from "@storybook/react";
 import { withA11y } from "@storybook/addon-a11y";
 
-import { Icon } from "@nimbus-ds/components";
+import { Icon, Link } from "@nimbus-ds/components";
 import { ExternalLinkIcon } from "@nimbus-ds/icons";
 
 import { HelpLink } from "./HelpLink";
@@ -14,22 +14,25 @@ export default {
     withA11y: { decorators: [withA11y] },
   },
   argTypes: {
-    link: { control: { disable: true } },
+    children: { control: { disable: true } },
   },
 } as ComponentMeta<typeof HelpLink>;
 
 const Template: ComponentStory<typeof HelpLink> = (args) => (
-  <HelpLink
-    {...args}
-    link={{
-      children: (
-        <>
-          Help link <Icon color="currentColor" source={<ExternalLinkIcon />} />
-        </>
-      )
-    }}
-  />
+  <HelpLink {...args} />
 );
 
 export const base = Template.bind({});
-base.args = {};
+base.args = {
+  children: (
+    <Link
+      as="a"
+      href="https://nimbus.tiendanube.com/"
+      target="_blank"
+      appearance="primary"
+      textDecoration="none"
+    >
+      Help link <Icon color="currentColor" source={<ExternalLinkIcon />} />
+    </Link>
+  ),
+};